# openapi.yaml
openapi: 3.0.0
info:
  title: 'Playin API'
  version: '0.5'
servers:
  -
    url: 'http://localhost'
    description: 'Local development environment'
  -
    url: 'https://api-dev.play-in.com'
    description: 'Development environment'
  -
    url: 'https://api-preprod.play-in.com'
    description: 'Staging environment'
  -
    url: 'https://api.play-in.com'
    description: 'Production environment'
paths:
...
  /api/events:
    get:
      tags:
        - Event
      summary: 'Get collection of events'
      operationId: 7058ab7c89d20c752b6afd6d808f6526
      responses:
        '200':
          description: 'A paginated collection of events'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPaginatedCollection'
    post:
      tags:
        - Event
      summary: 'Create event'
      operationId: 508895e03d7f10edac84357084c87b20
      requestBody:
        $ref: '#/components/requestBodies/EventCreateRequest'
      responses:
        '201':
          description: 'The created event'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Event' }
                type: object
        '400':
          description: 'Input format is incorrect'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 'Input data has not been validated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/api/stores/{store}/upcoming-events':
    get:
      tags:
        - Event
      summary: 'Get collection of upcoming events at a store'
      operationId: 383e8357a336d6fba59024bb75cf3aa0
      parameters:
        -
          name: store
          in: path
          description: 'The store where the upcoming events occur'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'A paginated collection of upcoming events at the given store'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPaginatedCollection'
  '/api/events/{id}':
    get:
      tags:
        - Event
      summary: 'Get event'
      operationId: 6f743f71f7a88dbe57be1d9c9ebf47cb
      parameters:
        -
          name: id
          in: path
          description: 'The ID of the event'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'The required event'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Event' }
                type: object
        '404':
          description: 'No event has been found with this ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Event
      summary: 'Update event'
      operationId: fb130dcf1ab4d7abbd3a173741df1f8f
      parameters:
        -
          name: id
          in: path
          description: 'The ID of the event'
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/EventCreateRequest'
      responses:
        '200':
          description: 'The updated event'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Event' }
                type: object
        '400':
          description: 'Input format is incorrect'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 'Input data has not been validated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/api/stores-credit-histories/{id}':
    get:
      tags:
        - 'Store Credit History'
      summary: 'Get store credit history entry'
      operationId: 16a72fccc14ff2153da790096b8435fa
      parameters:
        -
          name: id
          in: path
          description: 'The ID of the entry'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'The required entry'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/StoreCreditHistory' }
                type: object
        '404':
          description: 'No entry has been found with this ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/api/users/{user}/store-credit-history':
    get:
      tags:
        - 'Store Credit History'
      summary: 'Get user store credit history'
      operationId: d3ae22018e605f3213b130449bc1f928
      parameters:
        -
          name: user
          in: path
          description: 'The ID of the target user'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'A paginated collection of store credit history entries for the given user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreCreditHistoryPaginatedCollection'
...
components:
  schemas:
...
    EventEditInput:
      required:
        - name
        - store_id
      properties:
        name:
          description: 'The name of the event'
          type: string
          nullable: false
        store_id:
          description: 'The id of the managing store'
          type: integer
          nullable: false
        start_time:
          description: 'The time at which the event starts'
          type: string
          format: date-time
          nullable: false
        end_time:
          description: 'The time at which the event ends'
          type: string
          format: date-time
          nullable: false
        max_capacity:
          description: 'The maximum number of registrations'
          type: integer
          minimum: 0
          nullable: false
        price:
          description: 'The default registration price'
          type: number
          format: float
          nullable: false
      type: object
...
    Address:
      properties:
        id:
          description: 'The ID of the address'
          type: integer
          nullable: false
        address_name:
          description: 'A descriptive name of the address'
          type: string
          example: Home
          nullable: false
        recipient_name:
          description: 'The name of the person or service of the address'
          type: string
          example: 'John Doe'
          nullable: false
        street:
          description: 'The street address'
          type: string
          example: '1600 Amphitheatre Pkwy'
          nullable: false
        postal_code:
          description: 'The postal code'
          type: string
          example: '94043'
          nullable: false
        locality:
          description: 'The locality in which the street address is'
          type: string
          example: 'Mountain View'
          nullable: false
        country:
          description: 'The country'
          type: string
          example: USA
          nullable: false
        links:
          description: 'Links related to the address'
          properties:
            edit:
              description: 'Link to edit the address'
              type: string
              format: url
          type: object
          nullable: false
      type: object
...
    Event:
      required:
        - id
        - name
        - store
      properties:
        id:
          description: 'The ID of the event'
          type: integer
          nullable: false
        name:
          description: 'The name of the event'
          type: string
          example: 'PPTQ Modern'
          nullable: false
        store:
          $ref: '#/components/schemas/Store'
        start_time:
          description: 'The time at which the event starts'
          type: string
          format: date-time
          nullable: false
        end_time:
          description: 'The time at which the event ends'
          type: string
          format: date-time
          nullable: false
        max_capacity:
          description: 'The maximum number of registrations'
          type: integer
          minimum: 0
          nullable: false
        registrations_count:
          description: 'The current number of registrations'
          type: integer
          minimum: 0
          nullable: false
        price:
          description: 'The default registration price'
          type: number
          format: float
          minimum: 0
          nullable: false
        registrations:
          $ref: '#/components/schemas/Registration'
      type: object
...
    Registration:
      required:
        - id
        - user
        - event
        - paid
      properties:
        id:
          description: 'The ID of the registration'
          type: integer
          nullable: false
        user:
          $ref: '#/components/schemas/User'
        event:
          $ref: '#/components/schemas/Event'
        paid:
          description: 'Is the registration paid?'
          type: boolean
          nullable: false
      type: object
...
    StoreCreditHistory:
      required:
        - id
        - comment
        - credit
        - customer_id
        - collaborator_id
        - created_at
      properties:
        id:
          description: 'The ID of the store credit history entry'
          type: integer
          nullable: false
        comment:
          description: 'A comment describing the entry'
          type: string
          nullable: true
        credit:
          description: "The credit added to the customer's balance"
          type: number
          format: float
          nullable: false
        customer_id:
          description: 'The id of the customer tracked by the entry'
          type: integer
          nullable: false
        collaborator_id:
          description: 'The id of the collaborator responsible for the transaction'
          type: integer
          nullable: false
        created_at:
          description: 'The timestamp of the entry'
          type: string
          format: date-time
          nullable: false
      type: object
    Store:
      required:
        - data
      properties:
        id:
          description: 'The ID of the store'
          type: integer
          nullable: false
        name:
          description: 'The name of the store'
          type: string
          nullable: false
      type: object
    User:
      required:
        - id
        - name
        - email
      properties:
        id:
          description: 'The ID of the user'
          type: integer
          nullable: false
        name:
          description: 'The name of the user'
          type: string
          nullable: false
        email:
          description: 'The email address of the user'
          type: string
          format: email
          nullable: false
        addresses:
          description: 'The addresses linked to the user'
          type: array
          items:
            $ref: '#/components/schemas/Address'
          nullable: false
      type: object
    Error:
      required:
        - message
      properties:
        message:
          description: 'A message describing the error'
          type: string
          nullable: false
        errors:
          description: 'A list of validation errors'
          type: object
          nullable: false
      type: object
...
    EventPaginatedCollection:
      type: object
      allOf:
        -
          $ref: '#/components/schemas/PaginatedCollection'
        -
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Event'
          type: object
...
    PaginatedCollection:
      required:
        - data
      properties:
        links:
          required:
            - first
            - last
            - prev
            - next
          properties:
            first:
              type: string
              nullable: false
            last:
              type: string
              nullable: false
            prev:
              type: string
              nullable: true
            next:
              type: string
              nullable: true
          type: object
        meta:
          required:
            - current_page
            - last_page
            - from
            - to
            - per_page
            - total
            - links
            - path
          properties:
            current_page:
              type: integer
              nullable: false
            last_page:
              type: integer
              nullable: false
            from:
              type: integer
              nullable: false
            to:
              type: integer
              nullable: false
            per_page:
              type: integer
              nullable: false
            total:
              type: integer
              nullable: false
            links:
              type: array
              items:
                required:
                  - url
                  - label
                  - active
                properties:
                  url: { type: string, nullable: true }
                  label: { type: string, nullable: false }
                  active: { type: boolean, nullable: false }
                type: object
              nullable: false
            path:
              type: string
              nullable: false
          type: object
      type: object
...
    StoreCreditHistoryPaginatedCollection:
      type: object
      allOf:
        -
          $ref: '#/components/schemas/PaginatedCollection'
        -
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/StoreCreditHistory'
          type: object
...
  requestBodies:
...
    EventCreateRequest:
      description: 'Request body for creating or updating an event'
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EventEditInput'
...